dash_app/requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
psycopg2-binary==2.9.9
shap==0.46.0
matplotlib==3.8.4
